Switch from Kubernetes to standalone Docker image
This was a large change involving a significant rewrite of the
deployment script. Kubernetes wasn't really cutting it in terms of
performance for me (too many http connections and redirects seemed to
slow things down a bit), so I moved to my own server and set up NGINX on
that server as a simple Docker image.