Inital Gitea deployment
This commit adds a working Gitea deployment. It uses the base gitea
image provided by gitea with no modifications (for now). In order to
make this deployment I had to make some changes to the way Postgres and
NGINX get deployed (mainly network connection changes). I added a new
file in conf.d/ to handle connections to the git.ashtonc.ca domain.

Modifications remaining include disabling gravatar, adjusting the home
page, and changing the icon/favicon. In order to do this I may have to
create a custom Dockerfile that includes my customization files, but
I'll look into some other options too.

The biggest issue with the deployment is that SSH connections don't work
for now. I'm not certain this is an easy issue to resolve given that I'm
hiding behind Cloudflare.